// src/catalogs/zipper-catalog.generated.ts
export const ZIPPER_CONTRACTS_CATALOG: ZipperChainContractConfig[] = [
    {
        name: "Disperse",
        address: "0x5d9d358fFbacB5B29a3f427D28037D5E8D2B9D30",
        type: "",
        description: "",
        version: 1,
    },
    {
        name: "Multicall3",
        address: "0xF35A6AE5408fa1356064849D0BC3855f801aa6aC",
        type: "",
        description: "",
        version: 1,
    },
    {
        name: "accountDelegateRouter",
        address: "0x476cD16fbc5A03992698422bcfC6b695eA0f6e10",
        type: "",
        description: "Account Delegate Router contract",
        version: 1,
    },
    {
        name: "depositBatchRouter",
        address: "0x1Bb3892BD91e2b84Ab9ffd196061Fc0a170BB688",
        type: "",
        description: "Deposit Batch Router contract",
        version: 1,
    },
    {
        name: "fundingAccount",
        address: "0xBfF4F6224BC10f233dDB1E61E770d9832aabC7c4",
        type: "funding",
        description: "Funding Account contract",
        version: 1,
    },
    {
        name: "guardRegistry",
        address: "0xd71F60FD6f784Cc0aD8c25441568C48705D95f64",
        type: "",
        description: "Guard Registry contract",
        version: 1,
    },
    {
        name: "internalTransferFlare",
        address: "0x2E186b0b5b3b0aA496E573c28aCc83a8a8A7b75A",
        type: "",
        description: "Internal Transfer Flare contract",
        version: 1,
    },
    {
        name: "tradingGateway",
        address: "0xD3fecf5D39131e23b6B0f872cA0a21c8A5a30932",
        type: "trading",
        description: "Trading Gateway contract",
        version: 1,
    },
    {
        name: "validationBatchRouter",
        address: "0xb4c6e583fbbc511bEC625804FBdC72577348107D",
        type: "",
        description: "Validation Batch Router contract",
        version: 1,
    },
    {
        name: "zipperEndpoint",
        address: "0xae6B981BE9B73421eB1ba5372d1A4A937d63ffFB",
        type: "",
        description: "",
        version: 1,
    },
    {
        name: "zipperFactoryChain",
        address: "0x5E3ee633bb495442912940d88418950E406CfAf3",
        type: "",
        description: "",
        version: 1,
    },
    {
        name: "zipperFactoryFee",
        address: "0x8236C9bA6f36E7Ace4890Ad722b07564d63cB68C",
        type: "",
        description: "",
        version: 1,
    },
    {
        name: "zipperFactorySubVault",
        address: "0x512ec11246ABd36a2D9369850202E6e29113Fa10",
        type: "",
        description: "",
        version: 1,
    },
    {
        name: "zipperFactoryToken",
        address: "0x0CB99019aD0a9b51483BB4319DC017e9c1bC41D9",
        type: "",
        description: "",
        version: 1,
    },
    {
        name: "zipperFactoryVault",
        address: "0x417A033bC4f54b55a4D9bcF16e00fC6cdC53EAD7",
        type: "",
        description: "",
        version: 1,
    },
    {
        name: "zipperKey",
        address: "0x7Fa63d58F8C3221E7489168D2E9C7aC70718ed9c",
        type: "",
        description: "",
        version: 1,
    },
    {
        name: "zipperToken",
        address: "0x0CB99019aD0a9b51483BB4319DC017e9c1bC41D9",
        type: "",
        description: "",
        version: 1,
    },
];